VMS-36737: Qt6 API changes - incompatible with Qt5

+++ vms/client/nx_vms_client_desktop/src/ui/graphics/items/generic/masked_proxy_widget.cpp
@@ -24,9 +24,9 @@ void blitPixmap(const QPixmap& source, QPixmap& destination, const QPoint& locat
     blitter.drawPixmap(location, source);
 }
 
-int sceneDevicePixelRatio(const QGraphicsScene* scene)
+qreal sceneDevicePixelRatio(const QGraphicsScene* scene)
 {
-    int ratio = 1;
+    qreal ratio = 1;
     if (!scene)
         return ratio;
